Moved InhibitMidsParingPdu to interactions package. Correction to PDU content length.
